+++ contracts/token/ERC777/IERC777Recipient.sol
@@ -25,7 +25,7 @@ interface IERC777Recipient {
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes calldata userData,
         bytes calldata operatorData
     ) external;
